Move taxonomy posts query to file
