+++ internal/handlers/sources/handlers.go
@@ -188,7 +188,7 @@ func (s *Service) SourcePostsHandler(w http.ResponseWriter, r *http.Request) {
 
 	var (
 		err   error
-		posts models.Posts
+		posts *models.Posts
 	)
 
 	if data.IsCurrentUserAdmin() {
@@ -201,7 +201,7 @@ func (s *Service) SourcePostsHandler(w http.ResponseWriter, r *http.Request) {
 			s.rdb,
 			redisKey,
 			s.config.CacheTimeout,
-			func() (models.Posts, error) {
+			func() (*models.Posts, error) {
 				return s.postsRepo.GetSourcePosts(
 					r.Context(), sourceID, cursor, orderBy,
 				)
@@ -226,7 +226,7 @@ func (s *Service) SourcePostsHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	data.Posts = &posts
+	data.Posts = posts
 	if sourceID == "other" {
 		data.Posts.Title = "Other Uploads"
 	}
