+++ internal/repositories/posts/home.go
@@ -50,7 +50,7 @@ func (r *Repository) GetHomePosts(ctx context.Context, cursor, orderBy string) (
 		}
 	}
 
-	data := struct{ OrderByWhat, WhereCondition string }{order, where}
+	data := struct{ WhereCondition, OrderByWhat string }{where, order}
 	query, err := r.queryCache.Render("home_posts.sql", data)
 	if err != nil {
 		return nil, err
