+++ internal/repositories/posts/sql/faved_posts.sql
