+++ internal/repositories/posts/sql/sitemap_data.sql
@@ -73,3 +73,4 @@
 	FROM playlist
 
 	ORDER BY type, location;
+
